Added delay for entities to be generated in structures
This allows the structure to be built around the entity before the entity is in the area.

This way entities don't fall through blocks and possibly die.

In addition; minecarts will be slightly higher than scanned to allow them to fall onto a rail; without this buffer the minecarts will fall through the rail and the block below the rail.